chore(claude-sonnet-4-5): for main tasks add rename tasks to right click context me...
